% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{eval_criterion}
\alias{eval_criterion}
\title{Evaluate a row-wise criterion against a data frame}
\usage{
eval_criterion(criterion, data)
}
\description{
Evaluate a row-wise criterion against a data frame
}
\keyword{internal}
