% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{eval_group_criterion}
\alias{eval_group_criterion}
\title{Evaluate a group-level criterion — returns a logical vector (nrow(data))}
\usage{
eval_group_criterion(criterion, data)
}
\description{
Evaluate a group-level criterion — returns a logical vector (nrow(data))
}
\keyword{internal}
