% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{validate_hierarchy}
\alias{validate_hierarchy}
\title{Check that a hierarchy is compatible with a data frame}
\usage{
validate_hierarchy(hierarchy, data)
}
\arguments{
\item{hierarchy}{A \code{cf_hierarchy}.}

\item{data}{A data frame.}
}
\value{
\code{hierarchy}, invisibly, if all column names exist; otherwise errors.
}
\description{
Check that a hierarchy is compatible with a data frame
}
\keyword{internal}
